requirements.txt
torch~=1.13.0
GitPython
gitdb2
PyAutoGUI~=0.9.53
PyQt5~=5.15.7
PyQt5-sip
rlcard~=1.0.9
pywin32~=304
matplotlib~=3.6.1
opencv-python~=4.6.0.66

numpy~=1.23.4
Pillow~=9.2.0
requests~=2.28.1
logging~=0.4.9.6
six~=1.16.0
setuptools~=56.0.0
urllib3~=1.26.12
ujson~=5.5.0
uvicorn~=0.19.0
pydantic~=1.10.2
douzero~=1.1.0
Cython~=0.29.32
colorama~=0.4.6
fastapi~=0.85.1